Extract element functions into modules
