+++ .gitignore
@@ -17,3 +17,4 @@ yarn.lock
 # Editors
 .vscode/
 
+.DS_Store
