+++ src/scene/types.ts
@@ -1,0 +1,6 @@
+export type SceneState = {
+  scrollX: number;
+  scrollY: number;
+  // null indicates transparent bg
+  viewBackgroundColor: string | null;
+};
